Dev jacques defensive programming, change types from int to long long to enablelarge files processing.

+++ flextyper/includes/finder.h
@@ -68,7 +68,7 @@ public:
     ////////////////////////////////////////////////////////////////////////
     void searchMultipleIndexes(ResultsMap& indexPosResults, const KmerMap& kmerMap, const std::set<fs::path>& indexPaths,
                                const std::string& indexFileLocation, uint maxOccurences, bool parallel, uint threadNumber,
-                               bool printSearchTime, uint offset);
+                               bool printSearchTime, long long offset);
 
 
     ////////////////////////////////////////////////////////////////////////
@@ -90,7 +90,7 @@ public:
     /// \param threadNumber
     ////////////////////////////////////////////////////////////////////////
     void parallelSearch(ResultsMap& indexPosResults, const fs::path& indexFileLocation, const KmerMap& kmerMap,
-                        fs::path indexPath, uint maxOcc, uint threadNumber, bool printSearchTime, uint offset);
+                        fs::path indexPath, uint maxOcc, uint threadNumber, bool printSearchTime, long long offset);
 
     ////////////////////////////////////////////////////////////////////////
     /// \brief sequentialSearch searches for kmers inside a single FmIndex
@@ -102,7 +102,7 @@ public:
     /// \param maxOcc
     ////////////////////////////////////////////////////////////////////////
     void sequentialSearch(ResultsMap& indexPosResults, const fs::path& indexFileLocation, const KmerMap& kmerMap,
-                          fs::path indexPath, uint maxOcc, bool printSearchTime, uint offset);
+                          fs::path indexPath, uint maxOcc, bool printSearchTime, long long offset);
 
     ////////////////////////////////////////////////////////////////////////
     /// \brief multipleIndexesParallelSearch searches for kmers inside multiple
@@ -115,7 +115,7 @@ public:
     /// \param threadNumber
     ////////////////////////////////////////////////////////////////////////
     void multipleIndexesParallelSearch(ResultsMap &indexPosResults, const fs::path& indexFileLocation, const KmerMap& kmerMap,
-                                       const std::set<fs::path>& indexPath, uint maxOcc, uint threadNumber, bool printSearchTime, uint offset);
+                                       const std::set<fs::path>& indexPath, uint maxOcc, uint threadNumber, bool printSearchTime, long long offset);
 
     ////////////////////////////////////////////////////////////////////////
     /// \brief multipleIndexesSequentialSearch searches for kmers inside a multiple
@@ -127,7 +127,7 @@ public:
     /// \param maxOcc
     ////////////////////////////////////////////////////////////////////////
     void multipleIndexesSequentialSearch(ResultsMap& indexPosResults, const fs::path& indexFileLocation, const KmerMap& kmerMap,
-                                         std::set<fs::path> indexPath, uint maxOcc, bool printSearchTime, uint offset);
+                                         std::set<fs::path> indexPath, uint maxOcc, bool printSearchTime, long long offset);
 
 private:
     ////////////////////////////////////////////////////////////////////////
